Ecommerce_chatbot_AstraDB_langchain/requirements.txt
langchain-astradb
langchain
langchain-community
langchain-huggingface
langchain-google-genai
google-generativeai
datasets
pypdf
python-dotenv
flask

-e .